Add files via upload
In this project, a drawing of a Porsche 911 GT3 CUP 991 will be designed according to a particular dimension. The car has several views which we will be replicating in our group. I will be doing the left facing part of the car. First, we collected an image (given below) which I used as a source to determine the dimensions and then make a drawing. Finally, I created a PDF file which will contain the overall sketch of my part.